server/requirements.txt
pandas>=2.2
python-dotenv>=1.0
langchain>=0.2
langchain-core>=0.2
langchain-google-genai>=1.0
google-generativeai>=0.8
pydantic>=2.7
dateparser>=1.2
orjson>=3.10
diskcache>=5.6
loguru>=0.7
